+++ test/request-web-tests/room-tests.js
@@ -6,17 +6,13 @@ process.env.DISABLE_API_WEB_LISTEN = '1';
 
 const fixtureLoader = require('gitter-web-test-utils/lib/test-fixtures');
 const fixtureUtils = require('gitter-web-test-utils/lib/fixture-utils');
-const createUsers = require('gitter-web-test-utils/lib/create-users');
-const createGroups = require('gitter-web-test-utils/lib/create-groups');
 const assert = require('assert');
 const request = require('supertest');
 const env = require('gitter-web-env');
 const config = env.config;
-const logger = env.logger.get('request-web-tests:room-tests');
+const ensureMatrixFixtures = require('./utils/ensure-matrix-fixtures');
 const registerTestSynapseUser = require('./utils/register-test-synapse-user');
 
-const groupService = require('gitter-web-groups');
-const userService = require('gitter-web-users');
 const installBridge = require('gitter-web-matrix-bridge');
 const matrixBridge = require('gitter-web-matrix-bridge/lib/matrix-bridge');
 const GitterUtils = require('gitter-web-matrix-bridge/lib/gitter-utils');
@@ -26,8 +22,6 @@ const app = require('../../server/web');
 
 const serverName = config.get('matrix:bridge:serverName');
 const bridgePortFromConfig = config.get('matrix:bridge:applicationServicePort');
-const gitterBridgeBackingUsername = config.get('matrix:bridge:gitterBridgeBackingUsername');
-const gitterBridgeProfileUsername = config.get('matrix:bridge:gitterBridgeProfileUsername');
 
 // Finds the regex in the text and creates an excerpt so the test failure message can more easily be understood
 function findInText(text, regex, excerptBufferLength = 16) {
@@ -40,60 +34,6 @@ function findInText(text, regex, excerptBufferLength = 16) {
         Math.min(result.index + result[0].length + excerptBufferLength, text.length - 1)
       )
     };
-  }
-}
-
-async function ensureMatrixFixtures() {
-  const userFixtures = {};
-
-  // Create the backing bridge user on the Gitter side if it doesn't already exist.
-  // We don't have access to dependency inject this like we do in the smaller unit tests
-  // so let's just create the user like it exists for real.
-  const gitterBridgeBackingUser = await userService.findByUsername(gitterBridgeBackingUsername);
-  if (!gitterBridgeBackingUser) {
-    logger.info(
-      `Matrix gitterBridgeBackingUser not found, creating test fixture user (${gitterBridgeBackingUsername}) to smooth it over.`
-    );
-    userFixtures.userBridge1 = {
-      username: gitterBridgeBackingUsername
-    };
-  }
-
-  // Create the profile bridge user on the Gitter side if it doesn't already exist.
-  // We don't have access to dependency inject this like we do in the smaller unit tests
-  // so let's just create the user like it exists for real.
-  const gitterBridgeProfileUser = await userService.findByUsername(gitterBridgeProfileUsername);
-  if (
-    !gitterBridgeProfileUser &&
-    // Also make sure we're not trying to create the same user if they are configured to be the same
-    gitterBridgeProfileUsername !== gitterBridgeBackingUsername
-  ) {
-    logger.info(
-      `Matrix gitterBridgeProfileUser not found, creating test fixture user (${gitterBridgeProfileUsername}) to smooth it over.`
-    );
-    userFixtures.userBridgeProfile1 = {
-      username: gitterBridgeProfileUsername
-    };
-  }
-
-  // Re-using the test fixture setup functions
-  let f = {};
-  await createUsers(userFixtures, f);
-
-  const matrixDmGroup = await groupService.findByUri('matrix', { lean: true });
-  if (!matrixDmGroup) {
-    logger.info('Matrix DM group not found, creating test fixture group to smooth it over.');
-
-    // Re-using the test fixture setup functions
-    let f = {};
-    await createGroups(
-      {
-        groupMatrix: {
-          uri: 'matrix'
-        }
-      },
-      f
-    );
   }
 }
 
@@ -131,12 +71,17 @@ describe('Rooms', function() {
 
   describe('Matrix DMs', () => {
     let gitterUtils;
+    let stopBridge;
     before(async () => {
       await ensureMatrixFixtures();
 
-      await installBridge(bridgePortFromConfig + 1);
+      stopBridge = await installBridge(bridgePortFromConfig + 1);
 
       gitterUtils = new GitterUtils(matrixBridge);
+    });
+
+    after(async () => {
+      await stopBridge();
     });
 
     it(`Creates Matrix DM when visiting URL`, async () => {